requirements.txt
pytest>=7.0
pdf2image
Pillow
tqdm
google-generativeai
